Fix warning + some other logic tweaks in drive_mv()
Previously, was warning about "File started with multiple parents" even when file had NO parents, i.e. current_parents was NULL.

Also we now recognize renaming to the same name as a no-op.

Better reporting when adding a parent folder but not removing one, i.e. not a true 'move'.