Devote less attention to parent edge cases
In previous refactoring commit, I stopped checking whether both `path` and `parents` were specified in `drive_upload()`. I claim this is incredibly rare and not worth the code and headspace to maintain.

Let the behaviour be undefined / undescribed and, if it is wrong or surprising and actually matters, someone will speak up.