Introduce the general idea of overwrite NA/TRUE/FALSE (#266)
* Refactor to reveal similarity in these functions

* Devote less attention to parent edge cases

In previous refactoring commit, I stopped checking whether both `path` and `parents` were specified in `drive_upload()`. I claim this is incredibly rare and not worth the code and headspace to maintain.

Let the behaviour be undefined / undescribed and, if it is wrong or surprising and actually matters, someone will speak up.

* Don't try to delete a file more than once

Tests added in the next commit revealed that, for filepath input to `drive_rm()`, we tried to delete each file n times, where n = the number of filepaths for that file. n can be greater than 1 if a file has multiple parents.

* Test that drive_update() can add a parent folder

I am removing checks and tests around unusual parent situations elsewhere. I intend to send people to drive_update() for all weird parent folder situations. So I want an explicit test that you can add a parent folder this way.

* Centralize more path handling

* Use rationalize_path_name()

* Deprecate `parent`, in favor of `path`

Revisit this later with lifecycle

* Changes in drive_create() mean this test hits the API now

* Test rationalize_path_name()

* Introduce `overwrite`

Closes #230

* Oops forgot drive_rename()

* Add NEWS section

* Oops, forgot drive_mv()

* Test check_for_overwrite() and, therefore, the `overwrite` arguments

* No, really, `parent` is a file ID

Refine tests

* Show `overwrite` in examples

* Add handy "recency" example for drive_find()

* Tweak NEWS

* Extract overwrite_hits()

* Maybe it's a little aggressive/dangerous to empty trash here

But I have seen that the presence of files on Drive, in the trash, can affect tests. Tests should technically leave the trash the way the found it and be "trash-aware". But sometimes it's easiest to just clean it out.

* Add drive_put()

Truly closes #230

* Test for drive_put()

* More doc updates re: drive_put()

* Put in the usual skips

* Add "local" now that so many other functions deal in a "remote" overwrite