Convert `media` filepath to UTF-8 early
Motivated by solving the case where the user allows the Drive filename to be determined by the local filename, e.g. `basename(media)`

Ultimately, I couldn't get this to work for CJK (see notes left in tests), but there might be other names for which this is still helpful. It seems like a neutral-to-positive move.