Another improvement re: UTF-8 file names (#344)
* Introduce write_utf8()

Closes #342

* Fix scrub_filepath() on Windows

Provoked by:

PCRE pattern compilation error
	'PCRE2 does not support \F, \L, \l, \N{name}, \U, or \u'
	at 'sers\jenny\AppData\Local\Temp\RtmpCOHuRB\DESC-TEST-drive-download20c02191404b.txt'

I.e. the \U in `C:\Users\jenny\...` was causing this regex to fall over.

At this point, I can get by with a fixed string here, so I will.

* Convert `media` filepath to UTF-8 early

Motivated by solving the case where the user allows the Drive filename to be determined by the local filename, e.g. `basename(media)`

Ultimately, I couldn't get this to work for CJK (see notes left in tests), but there might be other names for which this is still helpful. It seems like a neutral-to-positive move.

* Add a test

* Add the usual skips