Import vctrs; bump min version of rlang and dplyr
We already depend on vctrs via tibble, so this isn't much of a change and makes it easier to add the vctrs compatibility layer.

I need recent rlang in order to get the `version` argument in `is_installed()`. I'm going to depend on bleeding edge rlang anyway due to the rlang + cli thing.

I'm implementing dplyr compatibility in a way that will only work for dplyr >= 1.0.0. I'm OK with this.