Update dplyr compat, add vctrs compat (#347)
* Import vctrs; bump min version of rlang and dplyr

We already depend on vctrs via tibble, so this isn't much of a change and makes it easier to add the vctrs compatibility layer.

I need recent rlang in order to get the `version` argument in `is_installed()`. I'm going to depend on bleeding edge rlang anyway due to the rlang + cli thing.

I'm implementing dplyr compatibility in a way that will only work for dplyr >= 1.0.0. I'm OK with this.

* Modernize dplyr compatibility; add vctrs compatibility

* Make it OK to reconstruct from list; tighten comments

* This concern is about 0-column case,not 0-row

* Add a tibble::tbl_sum method

* Put a minimum version on vctrs

* Add NEWS bullets

* Improve docs

* This is more inline with other tibble subclass implementations

* Use the new `expect_dribble()` elsewhere

* I've centralized these

* Makes more sense