devtools::build_readme()
The previous result is what tipped me off that the examples for `drive_create()` had gone off the rails and were not cleaning up after themselves.

That was not a problem with the code per se, but related to an API change that I adjusted to in b7e63826abcb632977aa8d8ffc84fdd116f9bde6 and c9a1a8be3f96fe27b0936212029429d70bd8eb95. Once the cleanup of a folder fails once, that can lead to additional errors, as happened here.